+++ assets/apps/agent/gestion_agents/models/index.ts
@@ -1,0 +1,5 @@
+import {Administration} from "./Administration.ts";
+import {Agent} from "./Agent.ts";
+import {RequeteAgentValidation, RequeteAgentValidationListe} from "./RequeteValidationAgent.ts"
+
+export {Administration, Agent, RequeteAgentValidation, RequeteAgentValidationListe}
