+++ tsconfig.json
@@ -6,6 +6,8 @@
     "jsx": "react",
     "esModuleInterop": true,
     "resolveJsonModule": true,
+    "experimentalDecorators": true,
+    "emitDecoratorMetadata": true,
     "types": ["vite/client", "vitest/globals"],
     "allowImportingTsExtensions": true,
      "baseUrl": ".",
